requirements.txt
fastapi
uvicorn[standard]
httpx
aiomysql
python-dotenv
